requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
argilla==1.1.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.3
colorclass==2.2.2
coloredlogs==15.0.1
compressed-rtf==1.0.6
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.8
Deprecated==1.2.14
diskcache==5.6.1
duckdb==0.8.1
easygui==0.98.3
ebcdic==1.1.1
et-xmlfile==1.1.0
extract-msg==0.41.5
fastapi==0.97.0
ffmpy==0.3.0
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gpt4all==0.3.4
gradio==3.35.2
gradio_client==0.2.7
h11==0.9.0
hnswlib==0.7.0
httpcore==0.11.1
httptools==0.5.0
httpx==0.15.5
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
IMAPClient==2.3.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.205
langchainplus-sdk==0.0.16
lark-parser==0.12.0
linkify-it-py==2.0.2
llama-cpp-python==0.1.64
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.0.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
olefile==0.46
oletools==0.60.1
onnxruntime==1.15.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.1
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandoc==2.3
pcodedmp==1.2.6
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
plumbum==1.8.2
ply==3.11
posthog==3.0.1
protobuf==4.23.3
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
pypandoc==1.11
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
red-black-tree-mod==1.20
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
RTFDE==0.0.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
unstructured==0.7.6
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.13.3
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zstandard==0.21.0
